py/requirements.txt
pandas
pymongo
openpyxl
pytz
python-dotenv